# PERF: CategoricalDtype.update_dtype