--- a/pandas/tests/dtypes/test_dtypes.py
+++ b/pandas/tests/dtypes/test_dtypes.py
@@ -222,7 +222,9 @@ class TestCategoricalDtype(Base):
 
     def test_update_dtype(self):
         # GH 27338
-        result = CategoricalDtype(["a"]).update_dtype(Categorical(["b"], ordered=True))
+        result = CategoricalDtype(["a"]).update_dtype(
+            CategoricalDtype(["b"], ordered=True)
+        )
         expected = CategoricalDtype(["b"], ordered=True)
         assert result == expected